+++ http/src/main/scala/com/twitter/finatra/http/filters/ExceptionMappingFilter.scala
@@ -3,12 +3,10 @@ package com.twitter.finatra.http.filters
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finagle.{Service, SimpleFilter}
 import com.twitter.finatra.http.internal.exceptions.ExceptionManager
-import com.twitter.util.Future
 import javax.inject.{Inject, Singleton}
 
 /**
  * Filter which converts exceptions into HTTP responses.
- *
  * NOTE: Should be as close to the start of the filter chain as possible.
  */
 @Singleton
@@ -17,11 +15,7 @@ class ExceptionMappingFilter[R <: Request] @Inject()(
   extends SimpleFilter[R, Response] {
 
   override def apply(request: R, service: Service[R, Response]) = {
-    (try service(request) catch {
-      case e: NoSuchMethodException =>
-        // Catch here instead of propagating it.
-        Future.exception(e)
-    }) handle { case e =>
+    service(request).handle { case e =>
       exceptionManager.toResponse(request, e)
     }
   }